Changes to show graph plot of test summary in the feeds
