+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/en-us.json
@@ -1150,13 +1150,13 @@
     "test-email-connection": "Test Email Connection",
     "test-entity": "Test {{entity}}",
     "test-plural": "Tests",
-    "test-result-summary": "Test Result Summary",
     "test-suite": "Test Suite",
     "test-suite-ingestion": "Test suite ingestion",
     "test-suite-lowercase": "test suite",
     "test-suite-lowercase-plural": "test suites",
     "test-suite-plural": "Test Suites",
     "test-suite-status": "Test Suite Status",
+    "test-suite-summary": "Test Suite Summary",
     "test-type": "Test type",
     "testing-connection": "Testing Connection",
     "tests-summary": "Tests Summary",
