+++ openmetadata-ui/src/main/resources/ui/src/utils/FeedUtils.tsx
@@ -42,13 +42,13 @@ import { EntityType, FqnPart, TabSpecificField } from '../enums/entity.enum';
 import { SearchIndex } from '../enums/search.enum';
 import {
   CardStyle,
+  EntityTestResultSummaryObject,
   FieldOperation,
   TestCaseStatus,
   Thread,
   ThreadType,
 } from '../generated/entity/feed/thread';
 import { User } from '../generated/entity/teams/user';
-import { TestCaseResult } from '../generated/tests/testCase';
 import {
   deletePostById,
   deleteThread,
@@ -658,11 +658,11 @@ export const getFieldOperationIcon = (fieldOperation?: FieldOperation) => {
 };
 
 export const getTestCaseNameListForResult = (
-  testResultSummary: Array<TestCaseResult>,
+  testResultSummary: Array<EntityTestResultSummaryObject>,
   status: TestCaseStatus
 ) =>
   testResultSummary.reduce((acc, curr) => {
-    if (curr.testCaseStatus === status) {
+    if (curr.status === status) {
       acc.push(curr.testCaseName ?? '');
     }
 
@@ -695,18 +695,19 @@ export const getTestStatusLabel = (status: TestCaseStatus) => {
   return statusLabelMapping[status];
 };
 
-export const formatTestStatusData = (testCaseResult: Array<TestCaseResult>) => {
-  const limitedData = testCaseResult?.slice(0, 3);
+export const formatTestStatusData = (
+  testResultSummary: Array<EntityTestResultSummaryObject>
+) => {
   const successCases = getTestCaseNameListForResult(
-    limitedData,
+    testResultSummary,
     TestCaseStatus.Success
   );
   const failedCases = getTestCaseNameListForResult(
-    limitedData,
+    testResultSummary,
     TestCaseStatus.Failed
   );
   const abortedCases = getTestCaseNameListForResult(
-    limitedData,
+    testResultSummary,
     TestCaseStatus.Aborted
   );
